+++ examples/prompting-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/fe-fpm-writer-ui
 
+## 0.0.10
+
+### Patch Changes
+
+-   Updated dependencies [ea0674c]
+    -   @sap-ux/ui-components@1.17.9
+    -   @sap-ux/ui-prompting@0.1.8
+
+## 0.0.9
+
+### Patch Changes
+
+-   Updated dependencies [b124873]
+    -   @sap-ux/ui-components@1.17.8
+    -   @sap-ux/ui-prompting@0.1.7
+
 ## 0.0.8
 
 ### Patch Changes