+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.40
+
+### Patch Changes
+
+-   Updated dependencies [d962ce1]
+    -   @sap-ux/project-access@1.27.1
+    -   @sap-ux/system-access@0.5.10
+
+## 0.0.39
+
+### Patch Changes
+
+-   Updated dependencies [df29368]
+    -   @sap-ux/project-access@1.27.0
+    -   @sap-ux/system-access@0.5.10
+
 ## 0.0.38
 
 ### Patch Changes