+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.3.19
+
+### Patch Changes
+
+-   Updated dependencies [d962ce1]
+    -   @sap-ux/project-access@1.27.1
+
+## 0.3.18
+
+### Patch Changes
+
+-   Updated dependencies [df29368]
+    -   @sap-ux/project-access@1.27.0
+
 ## 0.3.17
 
 ### Patch Changes